+++ NEWS.md
@@ -1,5 +1,6 @@
 # ggtibble 1.0.0.9000
 
+* `guides()` can now be added to `gglist` objects.
 * Labels created with the `labs` argument to `ggtibble()` will not longer all be
   the same (#3)
 * `new_gglist()` and `new_ggtibble()` are now exported making it easier to